Support for email attachments in multisite
This fix allows WP Broadbean to attach an uploaded CV to the application/contact email when WordPress is using multisite and the site being viewed is not the default site. Without this fix, the CV seems to get attached only when the file is uploaded for the first time. Subsequent uploads of the same file (or another file with the same name) results in the CV not getting attached.